In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,8 +10,8 @@
 
 int main(void)
 {
-	int sum, i;
-	int num1 = 0, num2 = 1;
+	long sum, i;
+	long num1 = 0, num2 = 1;
 
 	for (i = 0; i < 50; i++)
 	{
